move country_info into data submodule
