+++ nominatim/data/country_info.py
